backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.25.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
geopy==2.4.1
openrouteservice==2.3.3
qrcode[pil]==7.4.2
